教学班相关

+++ OnlineExamSystemApi/src/main/java/cn/edu/zjut/serviceimpl/TeachingClassServiceImpl.java
@@ -7,6 +7,7 @@ import cn.edu.zjut.service.TeachingClassService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -29,7 +30,14 @@ public class TeachingClassServiceImpl implements TeachingClassService {
     }
 
     @Override
-    public void deleteTeachingClass(Integer teachingClassId) {
+    public void deleteTeachingClass(Integer teachingClassId,String employeeNumber) {
+        List<TeachingClass> ownedClasses=teachingClassMapper.findByEmployeeNumber(employeeNumber);
+        boolean isOwned = ownedClasses.stream()
+                .anyMatch(classItem -> classItem.getTeachingClassId()==(teachingClassId));
+        if (!isOwned) {
+            throw new IllegalArgumentException("当前用户无权限删除该教学班，教学班ID: " + teachingClassId);
+        }
+
         // 删除教师与教学班的关联
         teachingClassMapper.deleteTeacherTeachingClass(null, teachingClassId); // 删除所有关联的教师
